+++ authentication/local.js
@@ -1,6 +1,6 @@
 module.exports = function(app, passport, provisioning) {
 
-  var log = require('winston')
+  const log = require('winston')
     , moment = require('moment')
     , LocalStrategy = require('passport-local').Strategy
     , User = require('../models/user')
@@ -10,26 +10,32 @@ module.exports = function(app, passport, provisioning) {
     , userTransformer = require('../transformers/user');
 
   function parseLoginMetadata(req, res, next) {
+    req.loginOptions = {
+      userAgent: req.headers['user-agent'],
+      appVersion: req.param('appVersion')
+    };
 
-    var options = {};
-    options.userAgent = req.headers['user-agent'];
-    options.appVersion = req.param('appVersion');
-
-    req.loginOptions = options;
     next();
   }
 
-  function authenticate(req, res, next) {
-    passport.authenticate('local', function(err, user, info) {
+  function isAuthenticated(req, res, next) {
+    if (!req.user) {
+      return res.sendStatus(401);
+    }
+
+    next();
+  }
+
+  function authorizeDevice(req, res, next) {
+    provisioning.provision.check(provisioning.strategy, {uid: req.param('uid')}, function(err, device) {
       if (err) return next(err);
 
-      info = info || {};
-      if (!user) {
-        return res.status(401).send(info.message);
+      if (provisioning.strategy === 'uid' && (!device || !device.registered)) {
+        return res.sendStatus(403);
+      } else {
+        req.device = device;
+        next();
       }
-
-      req.user = user;
-      next();
     })(req, res, next);
   }
 
@@ -40,12 +46,12 @@ module.exports = function(app, passport, provisioning) {
 
         if (!user) {
           log.warn('Failed login attempt: User with username ' + username + ' not found');
-          return done(null, false);
+          return done(null, false, {message: 'Please check your username and password and try again.'});
         }
 
         if (!user.active) {
           log.warn('Failed user login attempt: User ' + user.username + ' is not active');
-          return done(null, false);
+          return done(null, false, {message: 'Please check your username and password and try again.'});
         }
 
         if (!user.enabled) {
@@ -79,24 +85,76 @@ module.exports = function(app, passport, provisioning) {
     }
   ));
 
+  // DEPRECATED retain old routes as deprecated until next major version.
+  app.post(
+    '/api/login',
+    function authenticate(req, res, next) {
+      log.warn('DEPRECATED - The /api/login route will be removed in the next major version, please use /auth/local/signin');
 
-  // Create a new device
-  // Any authenticated user can create a new device, the registered field
-  // will be set to false.
+      passport.authenticate('local', function(err, user, info = {}) {
+        if (err) return next(err);
+
+        if (!user) {
+          return res.status(401).send(info.message);
+        }
+
+        req.user = user;
+        next();
+      })(req, res, next);
+    },
+    provisioning.provision.check(provisioning.strategy),
+    parseLoginMetadata,
+    function(req, res) {
+      new api.User().login(req.user,  req.provisionedDevice, req.loginOptions, function(err, token) {
+        res.json({
+          token: token.token,
+          expirationDate: token.expirationDate,
+          user: userTransformer.transform(req.user, {path: req.getRoot()}),
+          api: config.api
+        });
+      });
+    }
+  );
+
+  app.post(
+    '/auth/local/signin',
+    function authenticate(req, res, next) {
+      passport.authenticate('local', function(err, user, info = {}) {
+        if (err) return next(err);
+
+        if (!user) {
+          return res.status(401).send(info.message);
+        }
+
+        req.login(user, function(err) {
+          if (err) return next(err);
+
+          res.json({
+            user: userTransformer.transform(req.user, {path: req.getRoot()})
+          });
+        });
+      })(req, res, next);
+    }
+  );
+
+  // DEPRECATED retain old routes as deprecated until next major version.
   app.post('/api/devices',
     function authenticate(req, res, next) {
       passport.authenticate('local', function(err, user) {
         if (err) return next(err);
-        if (user) {
-          req.user = user;
-          next();
-        } else {
-          // skip to next route
+
+        if (!user) {
           return next('route');
         }
+
+        req.login(user, function(err) {
+          next(err);
+        });
       })(req, res, next);
     },
     function(req, res, next) {
+      log.warn('DEPRECATED - The /api/devices route will be removed in the next major version, please use /auth/local/devices');
+
       var newDevice = {
         uid: req.param('uid'),
         name: req.param('name'),
@@ -115,7 +173,37 @@ module.exports = function(app, passport, provisioning) {
           }
 
           Device.createDevice(newDevice)
-            .then(newDevice => res.json(newDevice))
+            .then(device => res.json(device))
+            .catch(err => next(err));
+        })
+        .catch(err => next(err));
+    }
+  );
+
+  // Create a new device
+  // Any authenticated user can create a new device, the registered field will be set to false.
+  app.post('/auth/local/devices',
+    isAuthenticated,
+    function(req, res, next) {
+      const newDevice = {
+        uid: req.param('uid'),
+        name: req.param('name'),
+        registered: false,
+        description: req.param('description'),
+        userAgent: req.headers['user-agent'],
+        appVersion: req.param('appVersion'),
+        userId: req.user.id
+      };
+
+      Device.getDeviceByUid(newDevice.uid)
+        .then(device => {
+          if (device) {
+            // already exists, do not register
+            return res.json(device);
+          }
+
+          Device.createDevice(newDevice)
+            .then(device => res.json(device))
             .catch(err => next(err));
         })
         .catch(err => next(err));
@@ -123,19 +211,24 @@ module.exports = function(app, passport, provisioning) {
   );
 
   app.post(
-    '/api/login',
-    authenticate,
-    provisioning.provision.check(provisioning.strategy),
+    '/auth/local/authorize',
+    isAuthenticated,
+    authorizeDevice,
     parseLoginMetadata,
-    function(req, res) {
+    function(req, res, next) {
       new api.User().login(req.user,  req.provisionedDevice, req.loginOptions, function(err, token) {
+        if (err) return next(err);
+
         res.json({
           token: token.token,
           expirationDate: token.expirationDate,
           user: userTransformer.transform(req.user, {path: req.getRoot()}),
+          device: req.device,
           api: config.api
         });
       });
+
+      req.session.destroy();
     }
   );
 };